Extract object-/owner-references back to structs, as reused in the core
