+++ kopf/toolkits/hierarchies.py
@@ -1,45 +1,8 @@
 """
 All the functions to properly build the object hierarchies.
 """
+from kopf.structs import bodies
 from kopf.structs import dicts
-
-
-def build_object_reference(body):
-    """
-    Construct an object reference for the events.
-
-    Keep in mind that some fields can be absent: e.g. ``namespace``
-    for cluster resources, or e.g. ``apiVersion`` for ``kind: Node``, etc.
-    """
-    ref = dict(
-        apiVersion=body.get('apiVersion'),
-        kind=body.get('kind'),
-        name=body.get('metadata', {}).get('name'),
-        uid=body.get('metadata', {}).get('uid'),
-        namespace=body.get('metadata', {}).get('namespace'),
-    )
-    return {key: val for key, val in ref.items() if val}
-
-
-def build_owner_reference(body):
-    """
-    Construct an owner reference object for the parent-children relationships.
-
-    The structure needed to link the children objects to the current object as a parent.
-    See https://kubernetes.io/docs/concepts/workloads/controllers/garbage-collection/
-
-    Keep in mind that some fields can be absent: e.g. ``namespace``
-    for cluster resources, or e.g. ``apiVersion`` for ``kind: Node``, etc.
-    """
-    ref = dict(
-        controller=True,
-        blockOwnerDeletion=True,
-        apiVersion=body.get('apiVersion'),
-        kind=body.get('kind'),
-        name=body.get('metadata', {}).get('name'),
-        uid=body.get('metadata', {}).get('uid'),
-    )
-    return {key: val for key, val in ref.items() if val}
 
 
 def append_owner_reference(objs, owner):
@@ -49,7 +12,7 @@ def append_owner_reference(objs, owner):
     Note: the owned objects are usually not the one being processed,
     so the whole body can be modified, no patches are needed.
     """
-    owner = build_owner_reference(owner)
+    owner = bodies.build_owner_reference(owner)
     for obj in dicts.walk(objs):
         refs = obj.setdefault('metadata', {}).setdefault('ownerReferences', [])
         matching = [ref for ref in refs if ref.get('uid') == owner.get('uid')]
@@ -64,7 +27,7 @@ def remove_owner_reference(objs, owner):
     Note: the owned objects are usually not the one being processed,
     so the whole body can be modified, no patches are needed.
     """
-    owner = build_owner_reference(owner)
+    owner = bodies.build_owner_reference(owner)
     for obj in dicts.walk(objs):
         refs = obj.setdefault('metadata', {}).setdefault('ownerReferences', [])
         matching = [ref for ref in refs if ref.get('uid') == owner.get('uid')]
